feat: move group select to shared folder
