+++ src/app/base/components/GroupSelect/GroupSelect.test.tsx
@@ -1,5 +1,6 @@
 import GroupSelect from "./GroupSelect";
 
+import { groupOptions } from "app/machines/constants";
 import { userEvent, render, screen } from "testing/utils";
 
 it("executes setGrouping and setHiddenGroups functions on change", async () => {
@@ -7,6 +8,7 @@ it("executes setGrouping and setHiddenGroups functions on change", async () => {
   const setHiddenGroups = jest.fn();
   render(
     <GroupSelect
+      groupOptions={groupOptions}
       grouping={null}
       setGrouping={setGrouping}
       setHiddenGroups={setHiddenGroups}
